Decompose reducer
