+++ src/index.js
@@ -3,11 +3,11 @@ import ReactDOM from "react-dom";
 import App from "./components/app/app.jsx";
 import {createStore, applyMiddleware} from "redux";
 import {Provider} from "react-redux";
-import {reducer} from "./reducer.js";
+import reducer from "./reducer/reducer.js";
 import thunk from "redux-thunk";
 import {composeWithDevTools} from "redux-devtools-extension";
 import {createAPI} from "./api.js";
-import {Operation} from "./reducer.js";
+import {Operation as DataOperation} from "./reducer/data/data.js";
 
 const api = createAPI();
 
@@ -16,12 +16,12 @@ const store = createStore(
     composeWithDevTools(applyMiddleware(thunk.withExtraArgument(api)))
 );
 
-store.dispatch(Operation.getMovies());
-store.dispatch(Operation.getPromoMovie());
+store.dispatch(DataOperation.getMovies());
+store.dispatch(DataOperation.getPromoMovie());
 
 ReactDOM.render(
     <Provider store={store}>
-      <App getComments={(movieId) => store.dispatch(Operation.getComments(movieId))} />
+      <App getComments={(movieId) => store.dispatch(DataOperation.getComments(movieId))} />
     </Provider>,
     document.querySelector(`#root`)
 );
