+++ src/reducer/namespace.js
@@ -1,0 +1,6 @@
+const Namespace = {
+  DATA: `DATA`,
+  STATE: `STATE`,
+};
+
+export default Namespace;
